resolved comments and moved _ordered private
